requirements.txt
absl-py==1.2.0
altair==4.2.0
altgraph==0.17.3
anyio==3.6.1
argcomplete==1.10.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.8.2
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
compressed-rtf==1.0.6
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
docx2txt==0.8
ebcdic==1.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.1
extract-msg==0.28.7
fastjsonschema==2.16.2
flatbuffers==22.9.24
fonttools==4.37.4
future==0.18.3
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
imageio==2.25.1
imageio-ffmpeg==0.4.8
IMAPClient==2.1.0
importlib-metadata==5.0.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.19.1
jupyter_client==7.3.5
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==2.0.4
moviepy==1.0.3
multimethod==1.9.1
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.6.1
nest-asyncio==1.5.6
networkx==3.0
nltk==3.7
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.3
oauthlib==3.2.1
olefile==0.46
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
pandas-profiling==3.6.6
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pdfminer.six==20191110
pefile==2022.5.30
phik==0.12.3
pickleshare==0.7.5
Pillow==9.2.0
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.7
pydeck==0.8.0b3
Pygments==2.13.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
Pympler==1.0.1
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.8
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
semver==2.13.0
Send2Trash==1.8.0
six==1.12.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
stack-data==0.5.1
statsmodels==0.13.5
streamlit==1.13.0
tangled-up-in-unicode==0.2.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminado==0.16.0
textract==1.6.5
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
visions==0.7.5
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xlrd==1.2.0
XlsxWriter==3.0.7
ydata-profiling==4.1.1
zipp==3.9.0
